+++ src/utils/_deepClone.ts
@@ -1,5 +1,7 @@
 import {
   isArray,
+  isBlob,
+  isFile,
   isFormData,
   isFunction,
   isMap,
@@ -11,7 +13,7 @@ import {
   isWeakSet,
   isWrapperObject,
 } from '../is'
-import { _cloneArray, _cloneArrayBuffer, _cloneFormData, _cloneMap, _cloneSet } from './_deepCloneBase'
+import { _cloneArray, _cloneArrayBuffer, _cloneBlob, _cloneFile, _cloneFormData, _cloneMap, _cloneSet } from './_deepCloneBase'
 
 export interface CustomCloner<T = any> {
   cloner(val: T, map: Map<any, any>): T
@@ -75,6 +77,13 @@ export function _deepClone<T>(obj: T, map: Map<any, any>, options: CloneOptions)
       obj.byteLength
     ) as T & DataView
     map.set(obj, res)
+  } else if (isFile(obj)) {
+    // File 继承自 Blob 所以需要先判断
+    res = _cloneFile(obj, map)
+    console.log('_cloneFile\n', obj, '\n', res);
+  } else if (isBlob(obj)) {
+    // Blob
+    res = _cloneBlob(obj, map)
   } else if (isWrapperObject(obj)) {
     res = Object(obj.valueOf())
     map.set(obj, res)